Remove some special handling for local files
Since we're now using the local file proxy, local files can be
treated just like normal URL files.

At first I tried leaving it using the local file function, but I wasn't
able to get it to consistently work because apparently it had a race
condition where sometimes the go button would be grayed out.

The state management with jquery was too complex for me to figure out
how to change safely to fix the gray issue. The best option seemed to be
removing the function entirely, since I haven't observed the race
condition with URLs yet, and to reduce the overall complexity of the
code by removing some.